# backend/internal/httpd/apispec/openapi.yaml
openapi: 3.1.0
info:
  description: Loopback-only HTTP surface served by the Go daemon. Generated from
    Go (code-first) — do not edit by hand; run `go generate ./...`.
  title: Agent Orchestrator HTTP daemon
  version: 0.1.0-route-shell
servers:
- description: Local daemon (loopback only)
  url: http://127.0.0.1:3001
paths:
...
  /api/v1/fleet:
    get:
      operationId: getFleetStatus
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ControllersFleetStatusResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Report the daemon-global fleet pause status
      tags:
      - projects
  /api/v1/fleet/pause:
    post:
      operationId: pauseFleet
      parameters:
      - description: Terminate live workers immediately instead of draining at idle.
        in: query
        name: hard
        schema:
          description: Terminate live workers immediately instead of draining at idle.
          type: boolean
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ControllersFleetStatusResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Pause the whole fleet; ?hard=true terminates live workers immediately
      tags:
      - projects
  /api/v1/fleet/resume:
    post:
      operationId: resumeFleet
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ControllersFleetStatusResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Resume the whole fleet
      tags:
      - projects
...
  /api/v1/projects:
    get:
      operationId: listProjects
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListProjectsResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: List all registered projects (active + degraded)
      tags:
      - projects
    post:
      operationId: addProject
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddProjectInput'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Register a new project from a git repository path
      tags:
      - projects
  /api/v1/projects/{id}:
    delete:
      operationId: removeProject
      parameters:
      - description: Project identifier (registry key).
        in: path
        name: id
        required: true
        schema:
          description: Project identifier (registry key).
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RemoveProjectResult'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Remove a project; stops sessions, cleans workspaces, unregisters
      tags:
      - projects
    get:
      operationId: getProject
      parameters:
      - description: Project identifier (registry key).
        in: path
        name: id
        required: true
        schema:
          description: Project identifier (registry key).
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectGetResponse'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Fetch one project; discriminates ok vs degraded
      tags:
      - projects
    put:
      operationId: updateProjectSettings
      parameters:
      - description: Project identifier (registry key).
        in: path
        name: id
        required: true
        schema:
          description: Project identifier (registry key).
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectSettingsInput'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Atomically replace a project's display name and config
      tags:
      - projects
  /api/v1/projects/{id}/config:
    put:
      operationId: setProjectConfig
      parameters:
      - description: Project identifier (registry key).
        in: path
        name: id
        required: true
        schema:
          description: Project identifier (registry key).
          type: string
      - description: Config ETag from the project read this replacement was built
          on. A stale token returns 409.
        in: header
        name: If-Match
        schema:
          description: Config ETag from the project read this replacement was built
            on. A stale token returns 409.
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetProjectConfigInput'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Replace a project's per-project config
      tags:
      - projects
  /api/v1/projects/{id}/pause:
    post:
      operationId: pauseProject
      parameters:
      - description: Project identifier (registry key).
        in: path
        name: id
        required: true
        schema:
          description: Project identifier (registry key).
          type: string
      - description: Terminate live workers immediately instead of draining at idle.
        in: query
        name: hard
        schema:
          description: Terminate live workers immediately instead of draining at idle.
          type: boolean
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Pause a project; ?hard=true terminates live workers immediately
      tags:
      - projects
  /api/v1/projects/{id}/resume:
    post:
      operationId: resumeProject
      parameters:
      - description: Project identifier (registry key).
        in: path
        name: id
        required: true
        schema:
          description: Project identifier (registry key).
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Resume a paused project
      tags:
      - projects
...
  /api/v1/projects/initialize:
    post:
      operationId: initializeProjectRepository
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InitializeRepositoryInput'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InitializeRepositoryResult'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Initialize a selected folder as a Git repository with an initial commit
      tags:
      - projects
...
components:
  schemas:
    APIError:
      properties:
        code:
          type: string
        details:
          additionalProperties: {}
          type: object
        error:
          type: string
        message:
          type: string
        requestId:
          type: string
      required:
      - error
      - code
      - message
      type: object
    AddProjectInput:
      properties:
        asWorkspace:
          type: boolean
        config:
          $ref: '#/components/schemas/ProjectConfig'
        name:
          type:
          - "null"
          - string
        path:
          type: string
        projectId:
          type:
          - "null"
          - string
      required:
      - path
      type: object
    AgentConfig:
      properties:
        effort:
          type: string
        model:
          type: string
        modelByHarness:
          additionalProperties:
            $ref: '#/components/schemas/DomainHarnessModel'
          type: object
        permissions:
          type: string
      type: object
...
    ControllersFleetStatusResponse:
      properties:
        paused:
          type: boolean
      required:
      - paused
      type: object
...
    DegradedProject:
      properties:
        id:
          type: string
        kind:
          enum:
          - single_repo
          - workspace
          - scratch
          type: string
        name:
          type: string
        path:
          type: string
        resolveError:
          type: string
      required:
      - id
      - name
      - kind
      - path
      - resolveError
      type: object
...
    DomainHarnessModel:
      properties:
        effort:
          type: string
        model:
          type: string
      type: object
    DomainModelAvailability:
      properties:
        harness:
          enum:
          - claude-code
          - codex
          - codex-fugu
          - aider
          - opencode
          - grok
          - droid
          - amp
          - agy
          - crush
          - cursor
          - qwen
          - copilot
          - goose
          - auggie
          - continue
          - devin
          - cline
          - kimi
          - kiro
          - kilocode
          - vibe
          - pi
          - autohand
          type: string
        message:
          type: string
        model:
          type: string
        observedAt:
          format: date-time
          type: string
        projectId:
          type: string
        reason:
          enum:
          - not-probed
          - no-capability
          - probe-unavailable
          - reachable
          - unreachable
          - recovered
          type: string
        status:
          enum:
          - unknown
          - reachable
          - unreachable
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
      - harness
      - model
      - status
      - reason
      type: object
...
    DomainReviewerConfig:
      properties:
        agentConfig:
          $ref: '#/components/schemas/AgentConfig'
        harness:
          type: string
      required:
      - harness
      type: object
    DomainWakeBackoffConfig:
      properties:
        base:
          description: Positive Go duration for the reset/base wake interval. Empty
            inherits wakeInterval.
          type: string
        enabled:
          description: When false, keep fixed-interval wake behavior at the base interval
            instead of exponential idle backoff. Defaults to true.
          type:
          - "null"
          - boolean
        max:
          description: Positive Go duration cap for exponential idle wake backoff.
            Empty uses the daemon default.
          type: string
      type: object
...
    InitializeRepositoryInput:
      properties:
        path:
          type: string
      required:
      - path
      type: object
    InitializeRepositoryResult:
      properties:
        path:
          type: string
      required:
      - path
      type: object
...
    ListProjectsResponse:
      properties:
        projects:
          items:
            $ref: '#/components/schemas/ProjectSummary'
          type: array
      required:
      - projects
      type: object
...
    Project:
      properties:
        agent:
          type: string
        config:
          $ref: '#/components/schemas/ProjectConfig'
        configETag:
          type: string
        defaultBranch:
          type: string
        drainingWorkers:
          type: integer
        id:
          type: string
        kind:
          enum:
          - single_repo
          - workspace
          - scratch
          type: string
        modelAvailability:
          items:
            $ref: '#/components/schemas/DomainModelAvailability'
          type: array
        name:
          type: string
        path:
          type: string
        pauseState:
          type: string
        paused:
          type: boolean
        repo:
          type: string
        workspaceRepos:
          items:
            $ref: '#/components/schemas/WorkspaceRepo'
          type: array
      required:
      - id
      - name
      - kind
      - path
      - repo
      - defaultBranch
      - paused
      type: object
    ProjectConfig:
      properties:
        agentConfig:
          $ref: '#/components/schemas/AgentConfig'
        agentRules:
          type: string
        agentRulesFile:
          type: string
        defaultBranch:
          type: string
        env:
          additionalProperties:
            type: string
          type: object
        maxLiveWorkers:
          type: integer
        orchestrator:
          $ref: '#/components/schemas/RoleOverride'
        orchestratorRules:
          type: string
        orchestratorRulesFile:
          type: string
        postCreate:
          items:
            type: string
          type: array
        prime:
          $ref: '#/components/schemas/RoleOverride'
        primeRules:
          type: string
        primeRulesFile:
          type: string
        reviewerRules:
          type: string
        reviewerRulesFile:
          type: string
        reviewers:
          items:
            $ref: '#/components/schemas/DomainReviewerConfig'
          type: array
        sessionPrefix:
          type: string
        symlinks:
          items:
            type: string
          type: array
        trackerIntake:
          $ref: '#/components/schemas/TrackerIntakeConfig'
        worker:
          $ref: '#/components/schemas/RoleOverride'
        workerMix:
          $ref: '#/components/schemas/WorkerMix'
      type: object
    ProjectGetResponse:
      properties:
        project:
          $ref: '#/components/schemas/ProjectOrDegraded'
        status:
          enum:
          - ok
          - degraded
          type: string
      required:
      - status
      - project
      type: object
    ProjectOrDegraded:
      oneOf:
      - $ref: '#/components/schemas/Project'
      - $ref: '#/components/schemas/DegradedProject'
      type: object
    ProjectResponse:
      properties:
        project:
          $ref: '#/components/schemas/Project'
      required:
      - project
      type: object
    ProjectSummary:
      properties:
        drainingWorkers:
          type: integer
        id:
          type: string
        kind:
          enum:
          - single_repo
          - workspace
          - scratch
          type: string
        name:
          type: string
        orchestratorAgent:
          type: string
        path:
          type: string
        pauseState:
          type: string
        paused:
          type: boolean
        resolveError:
          type: string
        sessionPrefix:
          type: string
      required:
      - id
      - name
      - path
      - kind
      - sessionPrefix
      - paused
      type: object
...
    RemoveProjectResult:
      properties:
        projectId:
          type: string
        removedStorageDir:
          type: boolean
      required:
      - projectId
      - removedStorageDir
      type: object
...
    RoleOverride:
      properties:
        agent:
          type: string
        agentConfig:
          $ref: '#/components/schemas/AgentConfig'
        wakeBackoff:
          $ref: '#/components/schemas/DomainWakeBackoffConfig'
        wakeInterval:
          description: Prime role only. Positive Go duration string such as 15m; empty
            uses the daemon default.
          type: string
      type: object
...
    SetProjectConfigInput:
      properties:
        config:
          $ref: '#/components/schemas/ProjectConfig'
      required:
      - config
      type: object
...
    TrackerIntakeConfig:
      properties:
        assignee:
          type: string
        enabled:
          type: boolean
        provider:
          enum:
          - github
          type: string
        repo:
          type: string
      type: object
...
    UpdateProjectSettingsInput:
      properties:
        config:
          $ref: '#/components/schemas/ProjectConfig'
        displayName:
          maxLength: 20
          minLength: 1
          type: string
      required:
      - displayName
      - config
      type: object
    WorkerMix:
      items:
        $ref: '#/components/schemas/WorkerMixEntry'
      type: array
    WorkerMixEntry:
      properties:
        agent:
          type: string
        effort:
          type: string
        model:
          type: string
        weight:
          type: integer
      required:
      - agent
      - weight
      type: object
...
    WorkspaceRepo:
      properties:
        name:
          type: string
        relativePath:
          type: string
        repo:
          type: string
      required:
      - name
      - relativePath
      - repo
      type: object
...